ReFact: Divided All Validation
