+++ app/seller/seller.router.js
@@ -1,7 +1,7 @@
 const router   = require( 'express' ).Router() ;
 
 const seller   = require( './seller.controller' ) ;
-const { validator } = require( '../../validator'     ) ;
+const { validator } = require( "../../middleware/validator" ) ;
 
 router.get( '/'     , validator.seller.search , seller.search ) ;
