+++ app/user/user.router.js
@@ -1,7 +1,7 @@
 const router = require( 'express' ).Router() ;
 
 const user = require( './user.controller' ) ;
-const { validator } = require( '../../validator' ) ;
+const { validator } = require( "../../middleware/validator" ) ;
 
 router.post( '/sign-up' , validator.user.signUp  , user.signUp  ) ;
 router.post( '/sign-in' , validator.user.signIn  , user.signIn  ) ;
