+++ middleware/validations/token.val.js
@@ -1,0 +1,8 @@
+const Joi = require( '@hapi/joi' ) ;
+
+const token = {
+    refreshToken : Joi.forbidden(),
+    accessToken  : Joi.forbidden(),
+} ;
+
+module.exports = { token };
